Update HomeScreen entry transition
* Reformat generated files

* Use `from-` & `toJson` from `HomeScreenArgument`

* Add variable type for readability

* Remove empty `date` folder

* Remove unused introduction_privacy_page

* Update data introduction golden

The copy was fixed from proof to prove.

* Update data incorrect goldens

* Update translations from Lokalise

* Use explicit typing

* Remove TestScreen and clean up HomeScreenArgs

* Clean up show HomeScreen api

* Make StackedWalletCards tappable

Navigate to the home screen when the stack is tapped.

* Run formatter

* Remove placeholderBuilder

* Prefer passing around WalletCard directly

Instead of passing only the cardFront to build screens we now pass
around the WalletCard so that all data is available, this is useful
for e.g. building the Hero widgets, where we want to rely on the
card's id as the hero tag for consistent behaviour.

* Clean up StackedWalletCards animation logic

* Tweak transition to home duration

* Add missing generated classes

* Make WalletCard JsonSerializable

To allow us to use the restorable navigation when passing the cards
around, the cards are now serializable. This way they can simply be
serialized to json before setting them as an argument.

* Animate cards onto dashboard

* Setup test animation

- Remove homescreen loading state when cards are provided
- Cache locked state in PinOverlay screen
- Wrap walletcard in hero widget
- Make wallet_card equatable (for bloc diff logic)

* Setup test env

Create a simple setup to show the personalization
success screen more easily, this can then be used
to test the dashboard transition.