Actually inject client dependencies into Wallet
Instead of having `Wallet` create both the `DigidAuthenticator` and
`PidRetriever` in its constructor, actually inject these dependencies by
passing them to the `init()` method. This also requires some integration
tests to be adjusted and some mock exports added.

This also adds `Wallet::create()` method that is specialized to using
the non-mock services, so that the wallet can still easily be created in
`init.rs`.