Add PkcePair and use it in modified openid client
This adds a `PkcePair` trait and implementing `S256PkcePair` struct,
which represents an PKCE code verifier and code challenge pair. This
effectively gets rid of the `pkce` dependency and expresses the
explicit relationship between the two in the type system.

This also modifies our additions to the `openid` crate to use this type.