Clean up wallet_app assets
* Remove leftover script

* Update pin_confirmation_failed.light golden

Copy was updated, update the golden to reflect this change.

* Fix --update-goldens

Make sure the goldens are saved into the correct folder when they are
updated.

* Update failing goldens

* Provide custom golden comparator

* Fix analyze issues

Ignore constant_identifier_names for WalletAssets file.

* Update ThemeScreen colors goldens

Updated the colors tab of the theme screen to use the onPrimary
color for the text as well, making sure it is properly readable.

* Update ColorStylesTab

Make it so that the text also uses the onPrimary
color so that it it actually readable.

* Reference all assets through WalletAssets

* Fetch assets through WalletAssets (wip)

* Define all assets in wallet_assets.dart

* Move forgot_pin_header to illustrations

* Move icon assets

Also remove ic_policy as it is available through Icons.policy

* Rename image_ prefixed images

* Move digid logo

* Moves illustrations to dedicated directory

* Moves logos to dedicated directory

* Move images to non-free directory

There was no clear distinction between images that were free and
non-free in our repository. Defaulting to having everything in the
non-free directory, we can move images that are explicitly marked as
free back to the normal images directory as needed.