Tweak wallet_app error handling
* Clean up wallet_mocks

Remove AppLifecycleService, as the service itself it trivial and
can be used directly, avoiding the need for a mock and fixing
the incompatibility with the dependency updates.

* Rename handleError method

* Formatting and extend ServerError

Extend ServerError on PinValidateServerError so that we can bubble
up any server errors to show the correct server error screen.

* Pull CheckPinResult error handling into bloc

Be more consistent in where (and how) we handle the FlutterApiErrors
that can occur, namely error them in the bloc using the dedicated
extension method. Also make CheckPinResult be a sealed class, so that
we can switch on it so that the CheckPinResultExtension can be deleted.

This changed required us to regenerate the mockito classes to keep
the tests working as expected.

* Improve FlutterApiError handling in Bloc