Migrate from DigidState to PidIssuanceEvent
Update the `wallet_app` code so that it support the new style of
deeplink processing, which now returns a PidIssuanceEvent instead of
a DigidState.

Also renamed all DigiD related classes to their PidIssuance equivalent
to keep the naming aligned with what happens and is returned from the
flutter api in `wallet_core`.