Introduce CoreError for typesafe error handling
Instead of relying on the FlutterApiError throughout the whole app this
commit introduces a (sealed) CoreError class, all FlutterApiErrors are
to be mapped into this class (or actually it's subclasses) through the
new CoreErrorMapper. This way we can exhaustively check all the error
types and add typesafe fields, used here to provide a RedirectError in
the CoreRedirectUriError.