Display No Internet ErrorScreen
* Run flutter pub upgrade

* Update nl copy

* Update WalletPersonalize Initial golden

Update the goldens since the copy was updated.

* Fix pin_bloc_test

Provide and configure mock for the checkHasInternet usecase so that the
correct networking error can be resolved. On creation it is also set to
BlocExtensions.checkHasInternetUseCase to align with the app's behavior.

* Update generic network error screen test

* Move instance to BlocExtension

Instead of keeping an instance in the usecase, opted
to move it into the BlocExtension class, where it will
only be used (at least for now). This should make it slightly
more clear why it isn't being injected like other usecases.

* Update NetworkError golden

* Add internet check

When a networking error occurs, verify if an active internet connection
is available and bubble the result up to the ui so that the correct
error screen can be displayed to the user.

* Introduce ErrorScreen.showNoInternet()

Add a method on the ErrorScreen class to show the new no internet page.

* Add 'no internet' error illustration

* Introduce CheckHasInternetUseCase

Introduce the CheckHasInternetUseCase, which can be used to reliably
determine if the user is connected to the internet.