Render pid preview exposed by core during pid issuance
* Implement MR Feedback

* Clean up PidAttributeMapper

Pull the formatting logic into the PidAttributeMapper and make the
core and data mappers solely responsible for providing the correct
attributes.

* Fix tests

* Make pid preview labels locale aware

* Remove mock knowledge from BloC

Simplify the bloc a bit by making the behaviour the same for mock
and non-mock build by providing all the needed preview attributes
at the _performMockDigidLogin() method directly.

* Revert unneeded const -> final conversions

At a point the Attributes could not be const due to validation, opted
to remove this and as such removing the implications of this change.

* Render Pid preview during personalization flow

This commit squashes changes from PVW-1166-display-pid-confirmation to
remove a config file that got committed, furthermore it:
- Removes the notion of AttributeType in one go
- Introduces the ActiveLocaleProvider to provide the app's active locale
to any component.
- Fixes a bug in the AutoLockObserver.
- Updates the Attribute and CoreAttribute to provide a localized label
in favor of bubbling all translations up through the whole app.
- Merges main into this branch.

* Fix flutter imports

* Split out PidAttributeMapper

Since we have a mock flow and the core (non mock) flow of getting the
pid attributes now, they also have to be mapped slightly differently.
Based on the current environment this commit now injects the relevant
mapper so that the previewAttributes can be transformed correctly.

Furthermore this also adds this missing translations in the mapper, as
we now provide the context to be able to do so.

* Pass on Pid Preview

Create a CoreAttribute model to represent the CardAttribute with the
needed type information on the flutter side and expand the
PidAttributeMapper so that it also properly handles this List of
CoreAttributes.