PVW-1169: Implement accept / reject pid functionality
* Log validation errors in pin_bloc

* Run dart fmt

* Fix broken tests

* Link TypedWalletCore to new reject api

* Support pid rejection

Prepare to notify the wallet_core when the user opts to reject
the offered pid

* Update WalletPersonalizeBloc for wallet_core

Update the WalletPersonalizeBloc so that it aligns better
with how the pid is issued in the wallet_core implementation.
This required some changes to keep the mock implementation working
as well, which is done by means of the AcceptOfferedPidUseCaseMock.

* Clean up instruction.rs

Instead of mapping all the potential 'Result<(), Error>' types to
WalletInstructionResult directly this commit introduces an extra
'TryFrom<InstructionError>' to avoid code duplication for this
common step of the conversion.

* Fix personalize flow back button

The app would crash when pressing back from the confirm issuance pin
screen because the attributes could not be resolved. This is fixed now
by making the attributes part of the current state so they can be
provided when the user presses back.

* Update WalletPersonalizeConfirmPinPage

Update the pin page inside the wallet personalisation
flow so that it calls through to the accept offered
pid usecase and handles those results.

* Update accept_pid_issuance to return WalletInstructionResult

* Remove invalid override annotation

* Refactor WalletUnlockResult -> WalletInstructionResult

Refactor it so it can be used more universally in places
that require the user to confirm with a pin.

* Introduce accept/reject usecases

* Add accept/reject methods to PidRepository