Update accept_pid_issuance returned errors
Update the accept_pid_issuance method so that it maps the nested
RemoteEcdsaKeyError into the corresponding PidIssuanceError.
This allows the nested InstructionError to be easily mapped into
a WalletInstructionError at the api.rs layer so that the ui can
in term show the matching UI (in this case the pin validation errors).