PVW-1312: Display cards on dashboard
* Update test formatting

* Update TypedWalletCore Tests

Add more tests to verify method calls are correctly
passed onto the wallet core and that the (expected)
FfiExceptions are converted to CoreErrors.

* Fix empty stream and duplicate card id

In the wallet core we now:
- Make sure to publish something to the card stream when
it is set, this should eventually be updated to publish
the actual cards. This makes sure we (in flutter_app) do
not have to deal with the stream being empty, which can
could occur when trying to grab the .first value of the
cards stream (causing the app to get stuck loading).
- The mock cards have been given a unique id, this makes
sure the hero animation works as expected, as it relies
on a unique id per card.

* Add containsPid() method

To redirect the user to the correct page on app start we need to know
whether the PID is already in the user's wallet, this commit adds this
method call to the WalletRepository and updates the corresponding
usecases so that the correct navigation path can be chosen on app
launch.

* Indefinitely observe core cards stream

Since the mdoc cards are not yet persisted the cards can
not be set on the stream at the moment the stream is set
through set_cards_stream(..). This means some events were
missed as the wallet_app did not start observing the cards
until personalisation flow is completed. To make sure the
wallet_app receives the correct (mock) state of the cards
we now observe the stream as soon as the wallet_core is
initialized.