PVW-1330: Update onboarding progress indicator
* Update goldens

Update the goldens with the new progress
indicator states.

* Update onboarding progress indicator

Instead of breaking up the progress indicator for
the privacy, setup pin and personalisation flows the
design has been updated to make it more like one big flow
as far as the progress indicator is concerned, this commit
implements those newly designed states.