PVW-1176: Implement special pin timeout flow
* Update TypedWalletCore test

* Refactor destroyWallet to resetWallet

Refactor notions (in impl and tests) from destroy
to reset wallet. Also updated failing tests and
make sure the core_wallet_repository calls through
to the wallet_core exposed reset api.

* Add setup failed golden tests

* Expose reset_wallet() core api

Add a (panicking) reset_wallet() method to the core
flutter api so that the wallet_app implementation of
the reset flow can be implemented.

* Navigate to ForgotPinScreen from TimeoutScreen

The timeout screen includes a 'forgot pin' cta that
(before this commit) navigated the user to the
PlaceholderScreen, since we already have a dedicated
ForgotPinScreen this has now been updated to navigate
there. To make the behaviour feel consistent (and not
have the app potentially navigate away from the forgot
pin screen when the timer expires) the _onTimeoutExpired
has also been slightly updated to only execute the
navigation when the TimeoutScreen is visible.

* Introduce ResetWalletUseCase

Add a usecase that can be used to reset the wallet
(only implemented for mock builds right now) and call
it from the forgot pin, wallet blocked and setup failed
screens.

* Introduce WalletPersonalizeSetupFailedScreen

Build the UI for the setup failed screen, which the user
should see when (s)he fails to enter her pin correctly
to confirm the offered pid. Instead of going into the normal
timeout flow we now show a dedicated screen (this setup failed
screen) since the user should really still remember his/her pin
at this point. The screen redirects the user into the setup flow
to reconfigure the wallet pin.