Map CardPersistence to cardId
The id field of the Card model has been updated
to reflect the persistence state, and only provide
a (non-null) id when the card is persisted. This
commit unwraps that id on the wallet_app side, and
simply default to an empty string when no id is
available, which should make sure the 'preview' flows
keep working without overcomplicating the model. Also,
good to know is that, currently, the id is only used as
a unique 'tag' for the hero animations.

Tests have also been updated to instantiate the new
persistence class.