PVW-1314: Update Card Mappers
* Update address subtitle mapper

Make the mapper look for an attribute
with the key 'city' to match the expected
key structure of the mdoc converted cards.

* Support core docTypes in subtitleMapper

Update the Subtitle mapper to not only support
the mock cards provided by the wallet_core but
also the mdoc converted cards.

* Make residential related keys explicit

Make the field names more explicit about the fact
that we are looking up the user's residential address.

* Simplify PidAttributeMapper

The mock and core implementations are now only
responsible for providing the correct keys to
perform the value lookups, simplifying their
implementations.

* Delete CoreAttribute class

The CoreAttributes were only used in the PidRepository and could be
easily replaced by the DataAttribute we use elsewhere throughout the
app, this also slightly simplifies the PidRepositoryImpl as it requires
less mapper classes to be injected. Finally it's a first step towards
simplifying the PidAttributeMapper as that can now always rely on the
same type of Attribute (namely the DataAttribute).

* Streamline mapper naming

Move attribute related mappers to attribute sub directory
and rename the used references where it made sense (to me)
to do so.

* Improve pid preview mapper

Make the birthName optional to avoid error-ing out
of the flow when the 'family_name_birth' attribute is
not available.

Also update some label copy.

* Handle incorrect pid attributes

Wrap the pid mapper in a try catch block to gracefully handle the case
where not all pid attributes are available and thus the preview can't be
generated. It now notifies the bloc of the error so that the generic
login failed screen can be shown.

* Enrich WalletCardDetail model in core builds

Since we (currently) know for that the cards are
issued by the RvIG, we simply hardcode this data
to make sure the UI has something to render. A fixme
and error log have also been added to make sure this is
picked up properly later.

* Merge WalletCardRepositories

Pull the fetching logic up to the datasource so that
the repository can be simplified.

* Introduce SubtitleMapper

Add a mapper class that extracts a subtitle from the
card attributes to be displayed as the second line on
the wallet card.

Also explicitly add the collection package so it's easier
to auto import its methods.

* Update gender icon

Use a more generic icon since it's not dynamic for now.

* Fix loading error in detail screen

Implement a CoreDataSource and update the
get_wallet_card_detail_usecase_impl to get the detail
screen to be able to load (some) data for core builds.

This also meant the DataAttributeRepository (sort of)
works fore core builds and thus the get_first_names_usecase_impl
could be simplified a bit again.

* Update get_first_names_usecase_impl

Update it to also support the mock (core) pid name attribute.

* Localise CardMapper

Add card title localisations to lokalise, opted
to remove the subtitle for now since there is no
clear strategy to lokalise it.

* Update CardValueMapper

Update the card value mapper to take the locale into
account and return string values based on the input
type, locale and app localisations.

* Fix incorrect label in pid attribute mapper

* Avoid menu crash

Update the get_first_names_usecase_impl.dart so that it
doesn't throw when running a 'core' build of the app, but
instead perform a naive firstname lookup.