PVW-1442: Improve SVG performance
* Remove .vec files

While testing it turned out that preloading
the .svg files resulted in a better user
experience than loading the .vec files at
runtime. Attempts were also made to preload
the .vec files but that did not work as expected,
making precaching the .svg files the better option
for the end user. For now we only precache the
pid related svgs (which happens to be all SVGs at this
time) to make sure the loading of the SVGs doesn't
take too long.

* Preload Pid SVGs

On the app start (while showing the splash screen)
we now attempt to preload the .svg files so they
can instantly be rendered as card backgrounds when
the WalletCardItems are shown.

* Update SvgOrImage for .vec support

* Precompile SVGs

Used the command:
find . -type f -name "*.svg" | xargs -I {} vector_graphics_compiler -i {} -o {}.vec
in /wallet_app