Use MdocsMap only for mock storage
This makes issuance in `mdoc` return `Vec<MdocCopies>` instead of
`MdocsMap`, as this is what `Wallet` is interested in and can easily be
returned without additional processing.

The `MdocsMap` is still available in the `mdoc` crate when enabling the
`mock` feature and can be used for mocking storage of mdocs.