PVW-1262: Dashboard card hero transition
* Implement MR Feedback

* Improve card shuttle animation

- Make the 'show details' cta support fade out animation.
- Clean up the animationStatus listener with an extension
that automatically removes the statusListener as well.

* Test CardDetailScreenArgument

* Restructure CardDetailScreen

Change the setup of the CardDetailScreen so that it
renders the WalletCardItem inside a Hero widget with
the id as a tag and make it so that it renders the card
in the loading state too (if it is provided) so that
the hero animation can be played.

Also moved the logic of the shuttle builder into the
WalletCardItem so that it can be easily re-used whenever
we want to apply it as some extra pizazz for the hero
animation.

* Update CardDetailBloc to expose card while loading

To implement the hero animation we need to be able to
show the card as soon as we start rendering the screen,
the first step to achieve this is to update the bloc
and arguments so that the initial loading state can
provide the card, which is done in this commit.