PVW-1411: Migrate to app_links
* Implement MR Feedback

* Update AppLifecycleService Test

Update it to match the new seeded value

* Minor cleanup

- Fix typo
- Remove unnecessary when call

* Update CheckNavigationPrerequisitesUseCaseTest

* Add PerformPreNavigationActionsUseCaseImpl Test

* Add CheckNavigationPrerequisitesUseCase Test

* Refactor DeeplinkService

This commit rewrites a large part of the DeeplinkService,
aiming to make reading and testing it a bit less complex.

Most notably the stream is broken up in two flows from
quite early, one with the supported wallet_app deeplinks,
which now perform the pre navigation actions and check the
prerequisites through usecases, and another flow where
handling of the uri is delegated to the wallet core, which
done as soon as the wallet reaches the unlocked state (but
without any other possible prerequisites).

* Inject Applinks into DeeplinkService

Inject the Applinks dependency so that it can be
swapped out and remove the need of the _uriController
since applinks provides a similar stream already.

* Introduce constants and CheckIsDeepdiveUsecase

- Introduce constants to make sure we always
rely on the same deeplink/dive host and path
- Introduce the CheckIsDeepdiveUsecase to
more clearly indicate what happens during the
debounce logic

# Conflicts:
#	wallet_app/lib/src/data/service/deeplink_service.dart
#	wallet_app/lib/src/domain/usecase/deeplink/impl/decode_deeplink_usecase_impl.dart
#	wallet_app/lib/src/navigation/wallet_routes.dart

* Replace uni_links with app_links