Implement constructor for IssuerSigned type
This adds a static `sign()` constructor method to `IssuerSigned` to
facilitate construction of such a type from an `UnsignedMdoc` and key
material. This method is then used in the by the issuer `Session`, which
now creates an intermediary `IssuerSigned` as a basis for creating a
`SparseIssuerSigned` type, without additional memory allocations.

This commit also includes some improvements minor improvements by
reducing the amount of `clone()` invocations.

The added method can be used by test code as a means to create a valid
`Mdoc` type.