PVW-1403: Render Disclosure Request
* Implement Rust MR feedback

* Update line breaks

For consistency match the spacing used in other methods
of the api.rs file.

* Rename result variable

* Implement FlutterApiErrorFields for ParseError

* Implement MR formatting feedback

* Remove Disclosure.Error variant

Errors should not be propagates as a result
type, as such removing the Disclosure.Error
variant which in term also simplifies the
wallet_app code.

* Simplify continue_pid_issuance

Since we are no longer using a stream the errors
now always end up on the wallet_app side and thus
can be logged there.

* Add TryFrom for IdentifyUriResult

Clean up api.rs by moving the mapping logic
to uri.rs.

* Fix typo and minor formatting issues

* Extract StartDisclosureResult

* Fix deeplink_service_test mock issue

* Core DisclosureSuccess support

Also render the DisclosureSuccess state in core builds
after successful pin confirmation.

* Add Accept/Cancel disclosure support

Call through to the wallet_core to accept
or cancel a disclosure request.

* Expose accept/reject disclose apis

* Fix formatting issues

* Update documentation

* Propagate disclosure translations

* Introduce DisclosureRepository

Pull the mapping logic out of the DisclosureBloc
and into the usecase / repository level to be consistent
with other blocs and reduce the complexity inside the bloc
itself.

* Expose DisclosureMissingAttributes state

Expose the DisclosureMissingAttributes through
the core disclosure bloc.

* Implement basic _onOrganizationApproved

Implement an initial version of the
_onOrganizationApproved callback so the user can
navigate to the second page on a core build inside
the disclosure screen.

* Fix Core disclosure stop flow

* Split up DisclosureBloc

Instead of having a single DisclosureBloc
for both mock and core builds this commit
splits things up so there is a MockDisclosureBloc
and a CoreDisclosureBloc. This because the mock
build relies on the whole flow being there from the
start, while in actuality this will not be the case
for the core builds. To make it easier to build the
core build while keeping the mock functionality
up and running I opted to split them up likeso.

* Refactor DisclosureResult

Rename from DisclosureEvent to DisclosureResult

* Refactor UriTypeResult

Rename IdentifiedUryType to UriTypeResult

* Refactor start- to continuePidIssuance

Align with the naming (and actual behavior) in the
wallet_core.

* Remove process_uri api

Remove the generic process_uri api in favor
of having a dedicated continue_pid_issuance
and start_disclosure api, which expect that
the provided uri is intended for their
specific usecase.

* Expose identify_uri api

Update the flutter api so that is exposes an
identify_uri api, so that we can eventually
move to more dedicated processing apis.

* Fix tests

Fix tests that were previously broken while
refactoring the deeplink layer.

* Refactor NavigationRequest to sealed class

Make the NavigationRequest a sealed class so we
can (more simply) specify the requirements of
navigation requests and by extension handle the
navigation accordingly (e.g. push vs pushAndRemoveUntil)

* Update PidIssuance Uri Flow

Instead of processing the complete uri in one
go and having the information flow from the
deelinkservice into the correct part of the app
this commit is the first step towards pulling that
apart, meaning the deeplinkservice will only identify
the uri and navigate to the correct screen with the
uri as the argument. The actual processing of the uri
is than the responsibility of the screen itself.

* Start DeeplinkService Simplification

So far the DeeplinkService has been responsible for
propagating all events that can come out of the
wallet_core based on the provided uri. This commit
is the first step towards simplifying that process,
where the deeplink service will only handle the
identification of the uri and opening the correct
screen which will in term process the uri itself.
Doing so will clean up the data flow as events will
no longer have to be propagated from the deeplinkservice
into the correct screen.