PVW-1568: Refactor wallet_app Attributes
* Fix formatting issues

* Update pubspec

* Make mock requested attributes explicit

Instead of relying on the MissingAttribute model in the
request objects used in the mock, setup a typealias with
some documentation to make it more explicit and clear of
how the object is used in these usecases.

* Update how we pass CardDetailScreenArgument

Use the toJson() method on CardDetailScreenArgument instead
of explicitly wrapping with jsonEncode and jsonDecode.

* Update Podfile.lock

* Update relevant goldens

Update the goldens that used to include the
profile picture (which is no longer available)

* Remove photo attribute assets

* Update LocalizedText tests

* Rename LocalizedLabels to LocalizedText

Refactor LocalizedLabels and extract the extensions
to a dedicated class file.

* Fix formatting

* Add PidAttributeMapper note

* Reorganize formatter & converters

Extract the LocalizedStringConverter & AttributeValueConverter and
create a AttributeValueFormatter to clean up the Attribute related
code.

* Update WalletPersonalizeBloc

Remove unused subscription, re-add //todo and move
state emission into finally block for better readability.

* Fix DataAttributeSection test

* Fix CardDataScreen test

* Fix CardSubtitleMapper test

* Fix CardFrontMapper test

* Fix card attribute mapper tests

* Refactor continuePidIssuance

Update the continuePidIssuance to be a Future instead
of a Stream and simplify the corresponding BLoC, this
is possible because the UI now takes care of the
translations and thus we don't need to propagate anything
new on a locale change.

* Fix history padding issue

Noticed the button was stuck to the bottom of the screen
because of some missing padding, since the change is minor
opted to add it here.

* Refactor StartDisclosure(uri)

Instead of returning a Stream, migrate to a simpler
Future as localisation is now handled by the UI directly,
making the Stream obsolete.

* Own LocalizedLabels

Instead of having LocalizedLabels extend from
LocalizedString which is 'owned' by wallet_core, we
have LocalizedLabels be a simple Map<String,String>,
this way we don't depend on the wallet_core
implementation as in our UI layer.

* Remove LocaleMapper completely

Remove the notion of the LocaleMapper, as actually selecting
the correct value to be used in the UI is now pulled up to the UI
layer, and thus selecting the correct value based on the active
locale in the Repository layer is no longer desired.

* Use Mapper<> where possible

Migrate from LocaleMapper<> to Mapper<> where the locale
is no longer needed.

* Introduce AttributeValue

Introduce the notion of an 'AttributeValue' to wallet_app, which
is very similar to the CardValue produced by the wallet_core and
thus is trivial to map from that. The benefits of AttributeValue
are that we own it, it's a sealed class so we can guarantee we
handle all cases, and we can extend it so callers can easily get
the localized, human readable value, finally it's typesafe in
relation to it's contents.

* Update Attribute to include Translations

Change Attribute class so that it now provides all
the translations instead of a single one, so it can
be passed on to the UI which can then decide on the
correct one.

* Remove AttributeValueType

As it's only used for mock builds, and not relevant for the MVP, opting
to remove the notion of AttributeValueType completely.