PVW-1561: Handle WalletState errors
* Print the stacktrace

Also print the stacktrace when a CoreStateError occurs
to supply a little more context.

* Refactor TypedWalletCore

Instead of having every api inserting its own
try/catch logic we now rely on a call() method
to call through the the WalletCore and handle
the errors from there.

* Handle WalletState error

Handle the WalletState error type by mapping it to a
CoreStateError and consecutively log it and crash the app when
it occurs as this error indicates a programming error that
should be fixed asap.