PVW-1418: Process QR contents using wallet_core
* Test CoreQrRepository

* Test CoreUriRepository

* Move business logic into UriRepository

Instead of having dedicated mock/core usecases
move the logic of processing URIs into the repository layer.

* Move business logic into QrRepository

Instead of having dedicated mock/core usecases
move the logic into the repository layer.

* Remove parsing logic from wallet_app

Instead of working with Uris in the wallet_app we now
treat any (potential) uri as a String and delegate all
the processing of that string to the wallet_core.

* Create dedicated screen arguments

Setup dedicated screen arguments for the
sign, issue and disclosure screens to be used when
navigating. Supporting both the mock flows and the
core flows.

* Make DecodeQrUseCase return NavigationRequest

Now it returns a NavigationRequest it can be split out
into a mock and core implementation, with the mock still
handling the QR codes from the demo while the core
implementation simply passes the contents of the qr on to
the core which in term tells the wallet_app what to do with
the scanned QR code.