PVW-1402: Render disclosure request in wallet_app
* Update LocalizedStrings import

Use the fully quilified path instead of importing the
LocalizedStrings to avoid ambiguity with LocalizedString.

* Update reader_auth.json

Country has been renamed to countryCode, updating the field here
to make sure it's provided correctly when generating the certs.

* Remove privacy_policy_url from Organization

* Reduce cloning in flutter_api disclosure models

* Resolve purpose based on activeLocale

* Avoid excessive cloning of reader_registration

* Add is_first_interaction_with_relying_party TODO

Make it explicit that this value still needs to be filled
with non-mock data at some point.

* Make policy_url conversion more rusty

* Remove RequestPolicy from RequestAttributesMissing

* Make data_storage_duration_days optional

On the flutter side the value was already converted
to a (secretly) optional storage duration, so that field
is now explicitly optional, meaning the section related
to the storage duration is hidden when unset. Kept the
i64 type as that is what duration.num_days() returns and
want to avoid things silently overflowing as a negligible
optimisation.

* Make countryCode explicit

To avoid any ambiguity when it comes to dealing with the
country field inside the Organization object, it is now
renamed to country_code, to make it explicit we are dealing
with a (ISO-3166-1 alpha-2) country code.

* Implement from for LocalizedString

Instead of using a trait to go from LocalizedStrings
to a Vec<LocalizedString> use a From implementation by
defining a wrapper struct (RPLocalizedStrings) that we
do own in this crate.

* Update sample reader_auth

- Add a kvk
- Update purpose statement

* Update COCOAPODS

* Simplify RelyingParty

Instead of having a RelyingPartyInfo struct
this commit aligns the RelyingParty model to be
more in line with the Organization model. It also
makes it explicit what (required) attributes are
still missing on the wallet_app side.

* Fix incorrectly displayed location label

* Refactor to .map over match

Instead of using the match statement to map to the
correct value, use the map function instead.

* Align organization naming (legal/display)

* Update wallet_app with for new RelyingParty

Update the wallet_app so that it supports the new
RelyingParty model which now includes more localized
strings.

* Update RelyingParty and From implementation

Update the RelyingParty model in wallet_core with newly introduced
fields and update the 'impl From<ReaderRegistration> for RelyingParty'
so that these fields are actually passed on to the wallet_app.

* Add missing RelyingParty fields

* Extract LocalizedText

Since the LocalizedText is used in other classes
besides the Attribute the typealias is now extracted
to a separate class.

* Introduce RelyingPartyInfo & Localizations

Make some of the RelyingParty(Info) fields translatable
and add the missing RelyingParty fields as part of the new
RelyingPartyInfo class.

* Maintain purpose onBackPressed

* Pass policy and purpose to the UI

Update the DisclosureResult model and the corresponding UI
state so that the policy and purpose coming from the core can
be rendered in the UI.

* Refactor hasPreviouslyInteractedWithOrganization

Rename hasPreviouslyInteractedWithOrganization to
isFirstInteractionWithOrganization so we use more consistent
naming everywhere and thus don't have to flip the boolean value
in the organization_detail_bloc.

* Pass organization to OrganizationDetailScreen

* Expand DisclosureResult model

Expand the DisclosureResult and RelyingParty models
to include more metadata as required by the UI.