PVW-1656: Format ISO-3166 Country codes
* Extract CountryCodeFormatter

Extract formatting logic of country code to country
label for re-usability.

* Add country_codes git ref docs

Add some documentation to explain why we are referring to
a specific commit while defining the dependency, including
a link to the MR we depend on to be merged before we can return
to the original package.

* Print localized country name

Refer to specific commit of the `country_codes` library which
includes the missing ISO lookup that we need to fetch the localized
country label through this library.

* Map CountryCode to localized country name

* Add kvk field to Organization

Add a dedicated KVK field and render it on the
OrganizationDetailScreen.

* Change RetentionPolicy max_duration type

Instead of using the 'Duration' type to propagate the maximum
storage duration, use a u64 and rename it to make it explicit
we are doing so in seconds.