PVW-1433: History mock API
* Update Issuance Organization field name

* Rename RelyingParty to Organization

In the flutter api we had a RelyingParty model but
since it's used both in issuance and disclosure it makes
more sense to rename this to Organization as well.

* Fix CardMapperTest

Verify all attributes are mapped using the updated mapList call.

* Remove Card from WalletEvent.Disclosure

Card is not needed as all required data is already part of the
RequestedCard model, as such remiving it from the model.

* Clean up (card) mapper

- Improve mapper api by using iterable, which is more versatile
- Improve RequestedCardMapper by building upon the normal CardMapper
so that we are not implementing the same thing twice

* Localize and map purpose

Localize and map the purpose string propagated
inside the WalletEvent object.

* Remove unused IssuanceStatus

For now the core will only support issued, so removing the
enum class for simplicity.

* Add missing history mock data

* Refactor Transaction to WalletEvent

Also add additional fields so the complete UI can be rendered
inside the wallet_app.

* Rename TransactionType values

* Rename date to date_time

Rename date to date_time to make it explicit that the
field is expected to also include a time timestamp.

* Fix api.rs history mock calls

Fix imports and instantiate new RP model

* Fix formatting issues

* Render Transactions in wallet_app

Update the wallet_app to map the transactions so they
can be rendered in the history overview screen.

* Call through to wallet_core

Add the history api calls to the TypedWalletCore
that expose the list of transactions, calling through
to the wallet_core.

* Expose mock history api

* Define basic history transaction model