PVW-1644: Move generated code to (dart) wallet_core package
* Add package pubspec.lock files

* Run dart formatter

* Update .gitattributes bridge_generated path

* Downgrade package sdk constraints

* Remove abstract TypedWalletCore

* Clean up documentation and unused files

* Provide mock/core WalletCore

When instantiating the TypedWalletCore provide the correct
WalletCore implementation based on the `provideMocks` flag.

* Remove generation packages from wallet_app

Packages that are now available through the wallet_core package
can be removed from the wallet_app's pubspec.yaml

* Setup wallet_mock package

Provide a mock implementation through the wallet_mock package.

* Move generated api to wallet_core package

Instead of having the generated code in the wallet_app directly,
this commit moves it into a wallet_core dart package to create
some separation of concerns and so we can eventually set up a
wallet_mock package to provide a mock implementation of the wallet
core api to reduce the complexity in wallet_core itself.