PVW-1648: Move mock business logic and data
* Fix old merge issue

Previously showing duplicate kvk, fixed by removing
one of the duplicate IconRows again.

* Move docType constants

Instead of defining the docTypes twice, move them
into the core/mock packages and reference them from
there.

* Disable driving license update flow

* Fix MR feedback

* Move TimelineAttributeRepository and add docs

* Process TimelineAttributes by docType

Make sure the docType is propagated correctly

* Simplify CoreQrRepository legacy logic

* Fix bug when scanning contract qr code

* Remove `provideMocks` from WalletUseCaseProvider

* Fix failing unit tests

* Fix lint issues

* Fix issue in core_history_repository

* Fix Card(Attribute)MapperTest

* Fix consistency issue in IssuanceState

The IssuanceCheckOrganization did not use named
parameters for it's parameters (as the only one),
so updated it to do so to be consistent with the
other states.

* Fix Running SignScreenTest

Golden tests are still failing for known reasons.

* Fix Running IssuanceScreenTest

Golden tests are still failing for known reasons.

* Remove MockWalletRepositoryTest

* Fix CorePidRepositoryTest

* Fix CheckIsValidPinUseCase

* Extract known docTypes

Extract the docTypes that are used for certain
business logic into the wallet_constants.dart.

Note that (to keep the loose coupling with mock/core)
these are also (redefined) in those respective packages.

* Extract CardConfig mapper

* Make docType on WalletCard required

* Regenerate generated classes

* Fix organization detail screen from history

* Fix 'cancellation not logged' bug

When disclosing for issuance the cancellation
was not logged correctly in all flows because
it wasn't triggered and/or the flag wasn't reset,
this is fixed with this commit.

* Remove more mock related classes

Remove the issuance_response, disclosure_request and mock_pid
repositories and all the usecases that depended on it that are
no longer needed. Also updated the renew driving license usecase
so that flow is working inside the mock again.

* Remove CoreSignRequestRepository

Remove the repository and update the WalletDependencyProvider so it
no longer relies on the provideMocks flag.

* Remove mock wallet & uri repository

* Clean up WalletDataSource

Remove unused methods and simplify implementations

* Remove organization datasource

Organizations are now solely fetched through (mock)
core so can be removed from the wallet_app.

* Repair History Details Screen

Update HistoryDetailsBloc and it's dependencies so that
the history details page can be rendered again.

* Call through to cancelIssuance

Call through to cancelIssuance to make sure the cancelled
disclose flow is also logged.

* Implement CardDetail through core

- Add a way to get issuer by docType
- Implement the readMostRecent apis from the
TimelineAttributes.

* Fix mock QR scanner

When scanning a edi qr code we now convert it to a
deeplink style uri that can be processed the same way
any other deeplink is processed.

* Support renewed license CardFront

* Avoid adding duplicate cards (by docType)

* Update CardFrontMapper

Update the card front mapper to support all the
cards that can be issued through the (mock) core
issuance flow. This logic should be moved to
the actual core at some point.

* Clean up IssuanceState

* Clean up SignState

Remove unused super and implement props

* Update navigation and add DocumentMapper

* Update uri recognition for Sign links

* Update sign_bloc

Update teh sign bloc so all data calls go through the
sign_repository and thus (mock) core.

* Introduce sign related usecases

* Update StartSigningResult with missing fields

* Explicitly import core models in repositories

In the repository layer explicitly import the core package
as 'core' so it's easier to see from and into which models
are mapped.

* Introduce SignRepository

* Introduce StartSignResult

* Update SignState and Sign Screen

Update these classes so the flow object can be removed.

* Move mock related data to mock directory

* Rename MockRequestedAttribute -> RequestedAttribute

* Introduce wallet_core_for_signing

* Support back navigation during issuance

* Support multi card issuance

* Support basic issuance

* Introduce DiscloseForIssuanceUseCase

The usecase to be used on the enter pin screen
inside the issuance flow.

* Update IssuanceBloc (wip)

Start updating the issuance bloc to fetch and process
data through the 'core'.

* Fix SetupMockedWalletUseCase

* Fix pid issuance mock issue

Also remove unused TimelineAttributeRepositoryImpl class

* Introduce and identify IssuanceNavigationRequest

* Complete CoreIssuanceRepository

* Setup Issuance WalletCore (mock) api

* Inject mock dependencies

* Improve (mock) EventLog api

* Improve (mock) Wallet api

* Setup disclosure and remove mock bloc

* Fix mock pid issuance flow

* Move issuance_responses to mock package

* Implement accept/reject disclosure

* Implement disclosure in mock

* Introduce wallet_event_log.dart

* AcceptPid through mock package

Configure the mock pid issuance flow so that it's a bit
closer to the core pid issuance flow by using a (const)
kMockPidIssuanceRedirectUri that is propagated to the
personalize screen and the correct (mock) issuance flow
is started based on that.

* Implement mock acceptPidIssuance

* Implement mock continuePidIssuance

* Introduce PinManager

Introduce a PinManager in the mock to simulate the
pin logic. Verify it's behavior with a test and use
it to implement the basics of the WalletCoreMock.

* Remove MockConfigurationRepository

Remove the configuration mock from the wallet_app
and implement it through the MockWalletCore.

* Implement basics of wallet_core_mock

Implement the locked flag, registration calls and
set/clear stream methods to allow the app to start
as expected.

* Add package pubspec.lock files

* Run dart formatter

* Update .gitattributes bridge_generated path

* Downgrade package sdk constraints

* Remove abstract TypedWalletCore

* Clean up documentation and unused files

* Provide mock/core WalletCore

When instantiating the TypedWalletCore provide the correct
WalletCore implementation based on the `provideMocks` flag.

* Remove generation packages from wallet_app

Packages that are now available through the wallet_core package
can be removed from the wallet_app's pubspec.yaml

* Setup wallet_mock package

Provide a mock implementation through the wallet_mock package.

* Move generated api to wallet_core package

Instead of having the generated code in the wallet_app directly,
this commit moves it into a wallet_core dart package to create
some separation of concerns and so we can eventually set up a
wallet_mock package to provide a mock implementation of the wallet
core api to reduce the complexity in wallet_core itself.